% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{biomass_con_1}
\alias{biomass_con_1}
\title{Equilibrium biomass for consumer
Type I Functional Response}
\usage{
biomass_con_1(m, a, e, r, K)
}
\description{
Equilibrium biomass for consumer
Type I Functional Response
}

