% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{biomass_res_1}
\alias{biomass_res_1}
\title{Equilibrium biomass for resource
Type I Functional Response}
\usage{
biomass_res_1(m, a, e)
}
\description{
Equilibrium biomass for resource
Type I Functional Response
}

