% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ode.R
\name{cr_ode_1}
\alias{cr_ode_1}
\title{CR model
Type I Functional Response}
\usage{
cr_ode_1(R0 = NULL, C0 = NULL, K = NULL, r = NULL, a = NULL,
  e = NULL, m = NULL, user = list(R0 = R0, C0 = C0, K = K, r = r, a = a, e
  = e, m = m))
}
\description{
CR model
Type I Functional Response
}

