% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ode.R
\name{cr_type_2}
\alias{cr_type_2}
\title{CR model
Type II Functional Response}
\usage{
cr_type_2(R0 = NULL, C0 = NULL, K = NULL, r = NULL, a = NULL,
  e = NULL, m = NULL, h = NULL, user = list(R0 = R0, C0 = C0, K = K, r =
  r, a = a, e = e, m = m, h = h))
}
\description{
CR model
Type II Functional Response
}

