% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{fc}
\alias{fc}
\title{Copy to destination}
\usage{
fc(fpath, tag = "-", des = "/home/souto1/Documents/")
}
\arguments{
\item{fpath}{File path of the source file}

\item{tag}{[optional] <chr> Tag to the filename}

\item{des}{Destination folder}
}
\description{
Copy file to destination and rename if desired.
}
