% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ft}
\alias{ft}
\title{flextable wrapper}
\usage{
ft(d, fnote = NULL, ttl = NULL)
}
\arguments{
\item{d}{A data frame}

\item{fnote}{[optional] <chr> Footnote}

\item{ttl}{[optional] <chr> Title}
}
\value{
A flextable object
}
\description{
Sugar function for default flextable output.
}
\examples{
mtcars |> ft()
}
