% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ggsrc}
\alias{ggsrc}
\title{Add source label to ggplot}
\usage{
ggsrc(plt, span = 2, size = 8, col = "grey55", lab = NULL, omit = "")
}
\arguments{
\item{plt}{A ggplot object}

\item{span}{<num> Number of lines: either 1 or 2}

\item{size}{<num> Text size}

\item{col}{<chr> Colour of the text}

\item{lab}{[optional] <chr> Custom label to use instead of the default}

\item{omit}{[optional] <chr> Text to omit from the label}
}
\value{
A ggplot object with the added label
}
\description{
Generate and add a source label with file path and run time to a ggplot object.
}
