% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{hcln}
\alias{hcln}
\title{Set colour scales}
\usage{
hcln(n, show = FALSE)
}
\arguments{
\item{n}{<num> Number of colours to output}

\item{show}{<lgl> `TRUE` to show the output colours}
}
\value{
Hex code of colours that can be used for plotting
}
\description{
Set colour scales for the desired number of colours.
}
\examples{
hcln(6,FALSE)
hcln(8,TRUE)
}
