% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{kbl2}
\alias{kbl2}
\title{kable wrapper}
\usage{
kbl2(d, fnote = NULL, cap = NULL)
}
\arguments{
\item{d}{A data frame}

\item{fnote}{[optional] <chr> Footnote}

\item{cap}{[optional] <chr> Caption}
}
\value{
A kable object
}
\description{
Sugar function for default kable output.
}
\examples{
mtcars |> kbl2()
}
