% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{label_src}
\alias{label_src}
\title{Generate source label}
\usage{
label_src(span = 2, omit = "")
}
\arguments{
\item{span}{<num> Number of lines: either 1 or 2}

\item{omit}{[optional] <chr> Text to omit from the label}
}
\value{
A label with source file path and run time
}
\description{
Generate a source label with file path and run time.
In interactive sessions, this function uses `rstudioapi` to get the file path.
It is designed to work in a script file in RStudio when running interactively.
It will return empty when running in the console.
}
\examples{
label_src(1)
}
