% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pars.R
\name{pars_mass}
\alias{pars_mass}
\title{Function for getting mass dependent paramaters
Constants from DeLong 2015}
\usage{
pars_mass(M)
}
\description{
Function for getting mass dependent paramaters
Constants from DeLong 2015
}

