% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pars.R
\name{pars_temp}
\alias{pars_temp}
\title{Function for getting temp dependent parameters
Standardized to some arbitrary value (here 15C)}
\usage{
pars_temp(t, t_stnd = 15)
}
\description{
Function for getting temp dependent parameters
Standardized to some arbitrary value (here 15C)
}

