% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{summ_cont}
\alias{summ_cont}
\title{Summary for continuous variables}
\usage{
summ_cont(d, cols, xname = "")
}
\arguments{
\item{d}{A data frame}

\item{cols}{Columns to summarise in a vector `c()` as unquote names or character}

\item{xname}{[optional] <chr> Character to omit in output column names}
}
\value{
A data frame of summarised variables
}
\description{
Summarise selected continuous variables.
Variables can be grouped beforehand.
}
\examples{
mtcars |> summ_cont(cols=mpg)
mtcars |> summ_cont(mpg,"mpg_")
mtcars |> group_by(vs,am) |> summ_cont(mpg,"mpg_")
mtcars |> group_by(vs,am) |> summ_cont(cols=c(mpg,disp))
}
