% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr-package.R
\docType{package}
\name{tsr-package}
\alias{tsr}
\alias{tsr-package}
\title{tsr: Convenient Functions Written by Tomas Sou}
\description{
Convenient functions written by Tomas Sou
}
\author{
\strong{Maintainer}: Tomas Sou \email{tomas.sou@carexer.com}

}
\keyword{internal}
