% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr-package.R
\docType{package}
\name{tsr-package}
\alias{tsr}
\alias{tsr-package}
\title{tsr: Convenient Functions Written by Tomas Sou}
\description{
Convenient functions written by Tomas Sou
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/soutomas/tsr}
  \item Report bugs at \url{https://github.com/soutomas/tsr/issues}
}

}
\author{
\strong{Maintainer}: Tomas Sou \email{tomas.sou@carexer.com}

}
\keyword{internal}
